Shared : move shared/database to server/database and compile
