+++ src/server/database/Database/QueryResult.cpp
