+++ src/server/database/Database/QueryResult.h
